test: link `nop` with `--as-needed`
On Ubuntu, gcc links by default with `--as-needed`, which prevents
any libraries specified on the command line from being linked into
the binary if they are not needed.

gcc on other Linux distributions, such as Fedora, or clangi do not
do this by default. Explicitly pass `--as-needed` for the `nop`
test binary for consistency.